requirements.txt
# Core dependencies for DINOv3 Alpha Matting
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer vision and image processing
opencv-python>=4.8.0
Pillow>=9.0.0
scikit-image>=0.21.0

# Data science and visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Scientific computing
scipy>=1.10.0

# Configuration and logging
PyYAML>=6.0
termcolor>=2.3.0

# Progress bars and utilities
tqdm>=4.65.0

# Optional: For advanced features
# tensorboard>=2.13.0  # For logging and visualization
# wandb>=0.15.0        # For experiment tracking
# albumentations>=1.3.0 # For advanced augmentations
